Refactored the logging infrastructure a bit
