--- TaskScheduler/Logging/IRedisConnectionFactory.cs
@@ -1,7 +1,0 @@
-﻿namespace TaskScheduler.Logging
-{
-    public interface IRedisConnectionFactory
-    {
-        IRedisConnectionWrapper GetConnection();
-    }
-}
