--- TaskScheduler/Logging/IRedisLogger.cs
@@ -1,7 +1,0 @@
-﻿namespace TaskScheduler.Logging
-{
-    public interface IRedisLogger
-    {
-        void Log<T>(T log) where T : LogStashLog;
-    }
-}
