+++ TaskScheduler/Logging/Messages/TaskPublishedLog.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using TaskScheduler.EventBus;
+
+namespace TaskScheduler.Logging.Messages
+{
+    public class TaskPublishedLog<T> : LogStashLog where T : IEvent
+    {
+        public T Event { get; set; }
+
+        public Type TypeOfEvent { get; set; }
+
+        protected override string Type
+        {
+            get { return "TaskPublishedLog"; }
+        }
+    }
+}
