+++ TaskScheduler/TaskScheduler.csproj
@@ -79,14 +79,11 @@
     <Compile Include="IDateTimeProvider.cs" />
     <Compile Include="IErrorLogRepository.cs" />
     <Compile Include="JSonConfigurationRepository.cs" />
-    <Compile Include="Logging\IRedisConnectionFactory.cs" />
-    <Compile Include="Logging\IRedisConnectionWrapper.cs" />
-    <Compile Include="Logging\IRedisLogger.cs" />
-    <Compile Include="Logging\LogStashLog.cs" />
+    <Compile Include="Logging\Messages\LogStashLog.cs" />
     <Compile Include="Logging\RedisConnectionFactory.cs" />
     <Compile Include="Logging\RedisConnectionWrapper.cs" />
     <Compile Include="Logging\RedisLogger.cs" />
-    <Compile Include="Logging\TaskPublishedLog.cs" />
+    <Compile Include="Logging\Messages\TaskPublishedLog.cs" />
     <Compile Include="Operations\EventOperation.cs" />
     <Compile Include="Operations\HttpOperation.cs" />
     <Compile Include="Operations\IOperation.cs" />
